Moved all common detail page pure functions to separate utilities file
