+++ src/pages/fishdetail.js
@@ -1,19 +1,6 @@
 import React, { useState, useEffect, useMemo } from "react";
 import { useParams } from "react-router";
-
-let monthMap = new Map();
-monthMap.set('1', 'January');
-monthMap.set('2', 'February');
-monthMap.set('3', 'March');
-monthMap.set('4', 'April');
-monthMap.set('5', 'May');
-monthMap.set('6', 'June');
-monthMap.set('7', 'July');
-monthMap.set('8', 'August');
-monthMap.set('9', 'September');
-monthMap.set('10', 'October');
-monthMap.set('11', 'November');
-monthMap.set('12', 'December');
+import { convertMonthNumToText, capitalizeFirstChar } from '../utilities/detailPageUtilities';
 
 /**
  * FishDetail
@@ -31,41 +18,6 @@ export default function FishDetail() {
         return await response.json();
     }
 
-    function monthNumToText(responseObj) {
-        let northernAvail =  responseObj["availability"]["month-northern"];
-        let southernAvail =  responseObj["availability"]["month-southern"];
-
-        let northernNumArr = northernAvail.split('-');
-        let northernText = monthMap.get(northernNumArr[0]) + "-" + monthMap.get(northernNumArr[1]);
-
-        let southernNumArr = southernAvail.split('-');
-        let southernText = monthMap.get(southernNumArr[0]) + "-" + monthMap.get(southernNumArr[1]);
-
-        // Update the key/values in the responseObj
-        responseObj = {
-            ...responseObj, 
-            "availability": { 
-                ...responseObj["availability"], 
-                "month-northern": northernText, 
-                "month-southern": southernText
-            }
-        }
-
-        return responseObj;
-    }
-
-    function capitalizeFirstChar(responseObj) {
-        let fishName = responseObj["name"]["name-USen"];
-
-        let nameArr = fishName.split(" ");
-        nameArr = nameArr.map((namePart) => {
-            return namePart.charAt(0).toUpperCase() + namePart.substring(1, namePart.length);
-        });
-
-        responseObj = {...responseObj, name: nameArr.join(" ")};
-        return responseObj;
-    }
-
     useEffect(() => {
         apiData
         .then((resp) => {
@@ -74,7 +26,7 @@ export default function FishDetail() {
 
             if (resp["availability"]["month-northern"] !== "" && resp["availability"]["month-southern"] !== "") {
                 // modify obj key to change the given month number to a text version (e.g. 1-5 becomes January-May),
-                resp = monthNumToText(resp);
+                resp = convertMonthNumToText(resp);
             }
                 
             updateFish(resp);
